requirements.txt
affine==2.4.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fiona==1.9.5
fonttools==4.49.0
GDAL==3.7.3
geopandas==0.14.3
geoserver-rest==2.6.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
rasterio==1.3.9
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
xmltodict==0.13.0
